dodanie klasy CheckForWinner
