+++ src/CheckForWinner.java
@@ -1,0 +1,143 @@
+import java.awt.*;
+
+public class CheckForWinner {
+    public static boolean checkForWinner(int currentColumn, int currentRow, Color colour) {
+        if (CheckForWinner.checkUpDownLeftRight(currentColumn, currentRow, colour)) return true;
+        if (CheckForWinner.checkDiagonally(currentColumn, currentRow, colour)) return true;
+        return false;
+    }
+
+    private static boolean checkDiagonally(int currentColumn, int currentRow, Color colour) {
+        int xStart;
+        int count;
+
+        int yStart;
+
+        //sprawdzenie lewej gory
+        count = 1;
+        yStart = currentRow;
+        xStart = currentColumn;
+        xStart--;
+        yStart--;
+        while(yStart>=0 && xStart>=0){
+            if(Drawer.grid[yStart][xStart].equals(colour)){
+                count++;
+            } else{
+                break;
+            }
+            if(count==4){
+                return true;
+            }
+            yStart--;
+            xStart--;
+        }
+
+        //sprawdzenie prawego dolu
+        count = 1;
+        yStart = currentRow;
+        yStart++;
+        xStart = currentColumn;
+        xStart++;
+        while(yStart< Drawer.grid.length && xStart< Drawer.grid[0].length){
+            if(Drawer.grid[yStart][xStart].equals(colour)){
+                count++;
+            } else{
+                break;
+            }
+            if(count==4){
+                return true;
+            }
+            yStart++;
+            xStart++;
+        }
+
+        //sprawdzenie lewego dolu
+        count = 1;
+        yStart = currentRow;
+        xStart = currentColumn;
+        xStart--;
+        yStart++;
+        while(yStart< Drawer.grid.length && xStart>=0){
+            if(Drawer.grid[yStart][xStart].equals(colour)){
+                count++;
+            } else{
+                break;
+            }
+            if(count==4){
+                return true;
+            }
+            yStart++;
+            xStart--;
+        }
+
+        //sprawdzenie prawej gory
+        count = 1;
+        yStart = currentRow;
+        yStart--;
+        xStart = currentColumn;
+        xStart++;
+        while(yStart>=0 && xStart< Drawer.grid[0].length){
+            if(Drawer.grid[yStart][xStart].equals(colour)){
+                count++;
+            } else{
+                break;
+            }
+            if(count==4){
+                return true;
+            }
+            yStart--;
+            xStart++;
+        }
+        return false;
+    }
+
+    private static boolean checkUpDownLeftRight(int currentColumn, int currentRow, Color colour) {
+        //sprawdzanie poziomo
+        int count = 0;
+        int xStart = currentColumn -3;
+        if(xStart<0){
+            xStart = 0;
+        }
+        int xEnd = currentColumn +3;
+        if(xEnd>6){
+            xEnd = 6;
+        }
+        while(xStart!=xEnd+1)
+        {
+            if(Drawer.grid[currentRow][xStart].equals(colour))
+                count++;
+            else{
+                count = 0;
+            }
+            if(count==4){
+                return true;
+            }
+            xStart++;
+        }
+
+        //sprawdzanie pionowo
+        count = 0;
+        int yStart = currentRow +3;
+        if(yStart>5){
+            yStart = 5;
+        }
+        int yEnd = currentRow -3;
+        if(yEnd<0){
+            yEnd = 0;
+        }
+        while(yStart!=yEnd-1)
+        {
+            if(Drawer.grid[yStart][currentColumn].equals(colour)){
+                count++;
+            }
+            else{
+                count = 0;
+            }
+            if(count==4){
+                return true;
+            }
+            yStart--;
+        }
+        return false;
+    }
+}
